share awaited spawn helper
Move runSpawn and SpawnExitError into a shared CLI helper and use it across create, init, and add so command actions await spawned processes consistently.

Also tighten scaffold selector typing by replacing filter(Boolean) with an explicit string predicate.
