+++ packages/cli/src/lib/run-spawn.ts
@@ -1,0 +1,32 @@
+import { spawn } from "cross-spawn";
+
+export class SpawnExitError extends Error {
+  code: number;
+
+  constructor(code: number) {
+    super(`Process exited with code ${code}`);
+    this.code = code;
+  }
+}
+
+export function runSpawn(
+  command: string,
+  args: string[],
+  cwd?: string,
+): Promise<void> {
+  return new Promise((resolve, reject) => {
+    const child = spawn(command, args, {
+      stdio: "inherit",
+      cwd,
+    });
+
+    child.on("error", (error) => reject(error));
+    child.on("close", (code) => {
+      if (code !== 0) {
+        reject(new SpawnExitError(code || 1));
+      } else {
+        resolve();
+      }
+    });
+  });
+}
